bias_analysis_service/requirements.txt
python-dotenv
google-generativeai
fastapi
uvicorn

